fastapi_auth/requirements.txt
fastapi==0.111.0
redis==5.0.4
gunicorn==23.0.0

sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
werkzeug==3.1.3

python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

alembic==1.14.0
typer==0.15.1

psycopg==3.2.3
psycopg-binary==3.2.3

opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-jaeger==1.21.0

user_agents==2.2.0

httpx[http2]==0.28.1
httpx-oauth==0.16.1